dom/workers/test/test_importScripts.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads (Bug 437152)
-->
<head>
  <title>Test for DOM Worker Threads (Bug 437152)</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=437152">DOM Worker Threads Bug 437152</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  var worker = new Worker("importScripts_worker.js");

  worker.onmessage = function(event) {
    switch (event.data) {
      case "started":
        worker.postMessage("stop");
        break;
      case "stopped":
        ok(true, "worker correctly stopped");
        SimpleTest.finish();
        break;
      default:
        ok(false, "Unexpected message:" + event.data);
        SimpleTest.finish();
    }
  };

  worker.onerror = function(event) {
    ok(false, "Worker had an error:" + event.message);
    SimpleTest.finish();
  }

  worker.postMessage("start");

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
